Move line_offset to a public attribute
